Determine activator/repressor arrow directions numerically, rather than symbolically
Previously, arrow directions were determined by symbolically testing whether kineticLaws
were monotonic functions. However, this approach fails for the Hill activator function,
as both the numerator and denominator are increasing functions of the activator concentration.

The new approach is to set all parameters to one, and compare the numerical value of the
kineticLaw when the concentration of the species of interest is 0 or 1. This fails to identify
non-monotonic effects, but these are less common.

The 'correct' approach of symbolically checking that the derivative of the kineticLaw w.r.t
the species of interest is always non-negative or non-positive is too complex for the limited
benefit it would provide.