Change the model abstraction process
If a species affects the rate of a reaction, the abstracted diagram contains an edge from it to every reactant and product of that reaction.
Previously, if a species was a reactant, no edges were added as a result of processing that edge.
Now, we skip only the edge from that reactant to itself (which would be redundant, as the rate of degredation of any species must be an
increasing function of its concentration).

For example, when processing the reaction S + I -> R, we would now add edges from S to R and from I to R; but, as before, we would not add
edges from S to S or from I to I.