Add option to represent transcription reaction with SBOlv glyphs
In this representation, the reaction node is replaced by a glyph of a promoter
followed by a CDS for each product. The promoter is assigned the reaction id,
so that regulatory arrows are drawn pointing at it. The product arrows are drawn
from the corresponding CDS.

A reaction is identified as transcription is if has an sboTerm attribute set
to SBO:0000183.