Visually indicate species with double-border if boundarySpecies is True (fill if it differs between models)
The rationale for this is that 'boundarySpecies' and 'constant' cause concentrations to not change how the
diagram would otherwise suggest.
We do not visualise 'constant' status, but this should not cause confusion because if a species has
constant=false but boundaryCondition=true, then the species cannot be a reactant or product (see Table 4
of L3V1 spec, in section 4.6.6).